WIP environ refatorada (builtins unset e export desativados)
