--- ms_temp_heredoc_0
@@ -1,2 +1,0 @@
-8
-s
